+++ openeo/rest/job.py
@@ -8,9 +8,9 @@ from typing import List, Union, Dict, Optional
 from deprecated.sphinx import deprecated
 from requests import ConnectionError, Response
 
-from openeo.rest import OpenEoClientException, JobFailedException
+from openeo.rest import OpenEoClientException, JobFailedException, OpenEoApiError
 from openeo.util import ensure_dir
-from openeo.internal.jupyter import render_component
+from openeo.internal.jupyter import render_component, render_error
 
 if hasattr(typing, 'TYPE_CHECKING') and typing.TYPE_CHECKING:
     # Only import this for type hinting purposes. Runtime import causes circular dependency issues.
@@ -266,8 +266,11 @@ class JobResults:
         return "<JobResults for job {j!r}>".format(j=self._job.job_id)
 
     def _repr_html_(self):
-        response = self.get_metadata()
-        return render_component("batch-job-result", data = response)
+        try:
+            response = self.get_metadata()
+            return render_component("batch-job-result", data = response)
+        except OpenEoApiError as error:
+            return render_error(error, "Job not finished yet")
 
     def get_metadata(self, force=False) -> dict:
         """Get batch job results metadata (parsed JSON)"""
